Add some vaccuous pass tests
Now, results with a value of 0 are valid, eg, it's valid to eat 0 carbs.
Include some other tests to cover the built in validator.  As it turns
out most of the record-specific validators preclude any invalid
elements, so the utils test ensures in the worst case that in the
general case no elements remain which have no valid and invalid date.

Side effect is that test coverage is now 100% of the
covered/instrumented sources.